move the rest to brain
