+++ docs/source/notebooks/3_generating_gaits.md
@@ -75,13 +75,13 @@ The parametric function is defined as follows:
 import jupyter_utils
 
 jupyter_utils.display_file(
-    'parametric_gait_generator.py',
+    '../../../packages/runtime/drqp_brain/drqp_brain/parametric_gait_generator.py',
     start_after='# Parametric function - START',
     end_before='# Parametric function - END',
 )
 ```
 
-```{literalinclude} parametric_gait_generator.py
+```{literalinclude} ../../../packages/runtime/drqp_brain/drqp_brain/parametric_gait_generator.py
 :start-after: '# Parametric function - START'
 :end-before: '# Parametric function - END'
 ```
@@ -92,13 +92,13 @@ With this function in place, we can now implement the gait generators for the th
 :tags: [remove-cell]
 
 jupyter_utils.display_file(
-    'parametric_gait_generator.py',
+    '../../../packages/runtime/drqp_brain/drqp_brain/parametric_gait_generator.py',
     start_after='# Gait params - START',
     end_before='# Gait params - END',
 )
 ```
 
-```{literalinclude} parametric_gait_generator.py
+```{literalinclude} ../../../packages/runtime/drqp_brain/drqp_brain/parametric_gait_generator.py
 :start-after: '# Gait params - START'
 :end-before: '# Gait params - END'
 ```
@@ -108,7 +108,7 @@ jupyter_utils.display_file(
 The `GaitsVisualizer` class provides a convenient way to visualize the gaits. The `visualize_continuous` method visualizes each axis of the gait in 2D, while the `visualize_continuous_in_3d` method visualizes the gait in 3D.
 
 ```{code-cell} ipython3
-from models import HexapodModel
+from drqp_brain.models import HexapodModel
 from parametric_gait_generator import GaitType, ParametricGaitGenerator
 from plotting import animate_hexapod_gait, GaitsVisualizer, is_sphinx_build
 
@@ -393,12 +393,12 @@ Now that we have all the pieces in place, we can put them together to create a f
 import jupyter_utils
 
 jupyter_utils.display_file(
-    'walk_controller.py',
+    '../../../packages/runtime/drqp_brain/drqp_brain/walk_controller.py',
     start_after='# THE SOFTWARE.',
 )
 ```
 
-```{literalinclude} walk_controller.py
+```{literalinclude} ../../../packages/runtime/drqp_brain/drqp_brain/walk_controller.py
 :start-after: '# THE SOFTWARE.'
 ```
 
@@ -406,9 +406,9 @@ jupyter_utils.display_file(
 import importlib
 
 import numpy as np
-from models import HexapodModel
+from drqp_brain.models import HexapodModel
 from plotting import animate_plot, is_sphinx_build
-import walk_controller
+import drqp_brain.walk_controller
 
 importlib.reload(walk_controller)  # autoreload fails with files written by notebook for some reason
