+++ packages/runtime/drqp_brain/drqp_brain/run_hexapod_ros.py
@@ -24,12 +24,12 @@ import argparse
 from enum import auto, Enum
 
 from drqp_brain.geometry import Point3D
-from models import HexapodModel
+from drqp_brain.models import HexapodModel
 import numpy as np
 import rclpy
 import rclpy.node
 import sensor_msgs.msg
-from walk_controller import GaitType, WalkController
+from drqp_brain.walk_controller import GaitType, WalkController
 
 kFemurOffsetAngle = -13.11
 kTibiaOffsetAngle = -32.9
